render-requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.6
email-validator==2.0.0
Werkzeug==2.2.3
pytesseract==0.3.10
Pillow==9.5.0
numpy==1.24.3
opencv-python==4.7.0.72
PyPDF2==3.0.1
python-docx==0.8.11
pdf2image==1.16.3
spacy==3.5.3